move toOffsetDateTime into Helpers and add user id check
